requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
prometheus_client==0.21.0
pytest==8.2.1
psycopg2-binary==2.9.9
httpx==0.27.2
pydantic-settings==2.4.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
langchain==0.2.11
langchain-community==0.2.5
chromadb==0.5.0
sentence-transformers==2.2.2
langchain-huggingface==0.0.3
